test: split admin logs tests
